requirements.txt
absl-py==1.4.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
astor==0.8.1
asttokens==2.2.1
attrs==21.4.0
autograd==1.5
autograd-gamma==0.5.0
ax-platform==0.2.10
backcall==0.2.0
beautifulsoup4==4.11.1
BioClients==0.2.8
bleach==5.0.1
botorch==0.8.0
cachetools==5.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chembl-webresource-client==0.10.8
cli-exit-tools==1.2.3.2
click==8.1.3
cloudpickle==2.2.0
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
DeepPurpose==0.1.5
defusedxml==0.7.1
descriptastorus @ git+https://github.com/bp-kelley/descriptastorus@86eedc60546abe6f59cdbcb12025a61157ba178d
dgllife==0.3.1
diff-match-patch==20200713
dill==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.10.0
django-import-export==3.0.1
djangorestframework==3.14.0
easydict==1.10
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
formulaic==0.5.2
fqdn==1.5.1
future==0.18.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gpytorch==1.9.0
grpcio==1.51.1
hyperopt==0.2.7
idna==3.4
interface-meta==1.3.0
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
lazy_loader==0.1rc2
lib-detect-testenv==2.0.3
lifelines==0.27.4
linear-operator==0.2.0
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.4
mock==5.0.1
multipledispatch==0.6.0
multiprocess==0.70.14
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy~=1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
pandas-flavor==0.3.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.12.0
prettytable==3.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.4
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.6
PyYAML==6.0
pyzmq==25.0.0
rdkit==2022.9.3
rdkit-pypi==2022.9.3
requests==2.28.1
requests-cache==0.7.5
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
subword-nmt==0.3.8
tablib==3.2.1
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typeguard==2.13.3
typing_extensions==4.4.0
uri-template==1.2.0
url-normalize==1.4.3
urllib3==1.23
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==3.6.1
xarray==2022.12.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
